Adds rerun version and prefix to module RPM dependencies
Update to stubbs:archive to fix #42 by adding the prefix "rerun-" to the list of REQUIRES specified in the module metadata file when building a module RPM. Additionally, the dependency on the rerun RPM is further narrowed to include the RERUN_VERSION used to build the RPM (e.g. "rerun=1.0").