Fixes options listing bug
Fixes #44 by making sure that options metadata is found relative to the resolved RERUN_MODULE_DIR for the specified module as opposed to incorrectly assuming the module is always going to be in RERUN_MODULES. This accommodates the case where a module (e.g. stubbs) is installed in /usr/lib/rerun/modules and used with /usr/bin/rerun and RERUN_MODULES is set to another location.